Fix a crash that happens when running C-GET.
Summary:
The problem happens when running sampleclient -get against the sampleserver. It
happens because a C-STORE subcommand issued in response to a C-GET picks the same message ID as the C-GET's.
This diff changes the message-ID-picking algorithm so that it actually looks at the list of used IDs and picks an unused one.

Reviewers: bwhitcher, jbredno

Reviewed By: jbredno

Differential Revision: https://phabricator.grailbio.com/D8233

fbshipit-source-id: 0cc862a